requirements.txt
numpy>=1.18.1
scipy
torch>=1.8.1
torchvision
tqdm
PyYAML
scikit-image
